### 2020-06-30 #2396 @lpantano
@lincoln-harris, can you make sure all the references has DOI, please?
