### 2020-06-30 #2396 @lpantano
@chrisleitzinger, I found your profile in GitHub and I thought you may be interested on reviewing this nice VCF reader tool. This are the guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Thanks!
