### 2020-07-03 #2432 @whedon
[REVIEW]: cerebra: A tool for fast and accurate summarizing of variant calling format (VCF) files
**Submitting author:** @lincoln-harris (<a href="http://orcid.org/0000-0003-2544-4225">Lincoln Harris</a>)
**Repository:** <a href="https://github.com/czbiohub/cerebra" target ="_blank">https://github.com/czbiohub/cerebra</a>
**Version:** v1.2.0
