### 2020-07-07 #2432 @lpantano
Hi @betteridiot, @afrubin!, this is the review issue, please accept the invitation if you haven't done it so you can start the review here.

@lincoln-harris, can you work on adding DOI to your references?

Thanks all!
