### 2020-07-12 czbiohub/cerebra#78 @afrubin
Versioning question
This repo has only a single release, 1.0.0. However, setup.py sets the version as 1.0.9 and the PyPI version is 1.1.0. `__init__.py` sets the version as 0.1.0.

Additionally, the seemingly-neglected HISTORY.md has only a single entry for 0.1.0 (despite this tag not existing in the repo).

This is important to reconcile as part of https://github.com/openjournals/joss-reviews/issues/2432
