### 2020-07-12 czbiohub/cerebra#79 @afrubin
Requirements and CI
It looks like conda_requirements.txt is a subset of requirements.txt and conda_requirements.txt is only used by the CI testing.

Why is this file included and why does the CI testing use conda at all? As a user who is looking to install the software, this is quite confusing.

requirements.txt also includes test-only requirements (coverage, flake8, etc.). These should be excluded from the requirements file and standard installation.

https://github.com/openjournals/joss-reviews/issues/2432
