### 2020-07-12 czbiohub/cerebra#80 @afrubin
Python 3.8 support
The software is not tested against Python 3.8, despite this being the current major release.

I suggest the authors add '3.8' '3.8-dev' and '3.9-dev' to the CI testing, with allowed failures for the dev versions. It also wouldn't hurt to test against (and allow failures for) 'pypy3'. Here is a simple example of this from one of my own projects: https://github.com/CountESS-Project/fqfa/blob/master/.travis.yml

Once 3.8 support is established, it should also be added to the setup.py classifiers section.

https://github.com/openjournals/joss-reviews/issues/2432
