### 2020-07-12 czbiohub/cerebra#81 @afrubin
Universal wheel
Currently the project has a setup.cfg file that specifies a universal wheel. Based on [the definition in the Python Packaging User Guide](https://packaging.python.org/guides/distributing-packages-using-setuptools/#universal-wheels) this is doesn't seem necessary or correct, since the project does not support Python 2 (nor should it at this point).

https://github.com/openjournals/joss-reviews/issues/2432
