### 2020-07-12 czbiohub/cerebra#82 @afrubin
Review comments on CONTRIBUTING.md
The CONTRIBUTING.md document includes most of the required material, but there are a few issues.

The links in the document lead to the site home pages rather than the specified pages due to misplaced `>`. The document would benefit from including additional links, such as a link to the relevant GitHub documentation in the first "Get Started!" list item.

This document has very different installation instructions to the README. They should be harmonized, and it will probably be simpler to maintain if CONTRIBUTING refers the user to the README, with notes about changes required to create a dev environment. It's quite confusing to have one document require conda and another require virtualenvwrapper. Ideally both should just use standard library venv.

This document specifies tox but the main project seems to use pytest. The Tips section uses unittest. The project should pick one testing framework and stick to it throughout.

https://github.com/openjournals/joss-reviews/issues/2432
