### 2020-07-14 #2432 @lincoln-harris
thanks for the comments, these are currently being addressed with https://github.com/czbiohub/cerebra/pull/84
