### 2020-07-15 czbiohub/cerebra#83 @lincoln-harris
as far as I can tell `pysam` just doesnt support windows: https://github.com/pysam-developers/pysam/issues/575. We dont directly use `pysam` -- its imported by a package that we call (though I cant figure out which one).

bc we're not directly using it its not a matter of converting to a pure python BAM parser like [`bamnostic`](https://github.com/betteridiot/bamnostic).

are there any potential workarounds to this?

tagging @betteridiot
