### 2020-07-16 czbiohub/cerebra#83 @afrubin
I think that if you clearly document that you don't support Windows installs due to the pysam dependency that should be sufficient. Explicitly stating the nature of the incompatibility is important in case pysam changes or the package that requires pysam changes in the future.

There are a lot of useful bioinformatics tools that won't run/compile on Windows, so most users are familiar with navigating this situation. [Windows Subsystem for Linux](https://docs.microsoft.com/en-us/windows/wsl/install-win10) should be a good workaround for users who want to run things locally, and very few academic HPC environments are Windows-based.
