### 2020-07-16 czbiohub/cerebra#82 @afrubin
This issue was closed as part of the PR, but was not actually resolved.
