### 2020-07-16 czbiohub/cerebra#82 @lincoln-harris
@lincoln-harris update https://github.com/czbiohub/cerebra/blob/master/CONTRIBUTING.md
