### 2020-07-16 czbiohub/cerebra#83 @betteridiot
I appreciate the tag @lincoln-harris . Attempting to be platform agnostic is a lofty goal when working in the genomics space. I see that you have removed `pysam` from your `requirements.txt`, but the dependency still exists in your `environment.yml` as well.

In developing `bamnostic`, I quickly realized how prevalent the norm of a linux-like environment is. It is perfectly valid to state that (as of now) your software works in Linux-based environments. This is because many of your dependencies are inextricably tied to linux-based software (`pathos`, `cython` (which is touch and go on Windows environments at the best of times), and even `colorama` can cause some issues).

If the intention of the package is to _analyze_ VCF files, then it is perfectly fine to rely on these other tools that have shown a long history of maintenance/support.

I agree with @afrubin that the expectation of a linux-like development environment is completely acceptable-**so long as you state as such**. An easy place to do this is in the `setup.py` using [classifiers](https://pypi.org/classifiers/) for your program.
