### 2020-07-16 czbiohub/cerebra#89 @betteridiot
JoSS Review: Clarify README.md
With regard to the ongoing review for your [JoSS submission](https://github.com/openjournals/joss-reviews/issues/2432), I would like to suggest stating some sort of use cases and/or warnings in your `README.md`. In particular, one of the largest drawbacks is the case of large VCF files. You even state as much in your [code](https://github.com/czbiohub/cerebra/blob/a6e4463fc6cff5b02583c8020cff09a288cbf598/cerebra/VCF.py#L61-L68):

>```python
> def dataframe(filename, large=True):
>     """ [...]
>     Note: Using large=False with large VCF files. It will be painfully slow.
>     [...]
>     """
> ```

This same cannot be said about your instructions/quickstart.

VCFs can be excruciatingly large sometimes (looking at you 1KG/ENCODE). Many times, researchers want/need to summarize from this starting point. Therefore, suggesting something like _filtering_ the VCF based on region/contig _prior_ to running `cerebra` would be thoughtful suggestion to gung-ho users to preempt potentially long-running jobs.
