### 2020-07-16 czbiohub/cerebra#90 @betteridiot
JoSS Review: Quality of writing
With regard to the "Quality of writing" section for the ongoing review for your JoSS submission (https://github.com/openjournals/joss-reviews/issues/2432), these are my (very minor/nitpicky) remarks:

* "wrangling" is somewhat parlance and may be a difficult word to discern context from for international audiences
* I may be wrong here, but surnames should be written for both in-text and bibliographic citation. For groups of 6 or more, it is acceptable to use "_et al_" after the first author. This simplifies the citation so as to not break up the thought of the sentence any more than it already has.
* You are already use LaTex to write the paper, so it may be simpler to write 10^8 as <img src="https://render.githubusercontent.com/render/math?math=10^{8}"> instead
* (Question) What happens if a `*.fai` is already provided prior to `find-peptide-variants`? Will it construct a new one?
* It might be helpful to state in the caption of **Figure 3** that the first 10+ minutes appear to be the construction of the `genome interval tree` and can be considered overhead for the process, which runs approximately linearly thereafter (Ha! Just saw that you state the same thing just below)
