### 2020-07-16 czbiohub/cerebra#90 @betteridiot
If you are not using LaTex, you can leverage GitHub rendering for your Markdown: `<img src="https://render.githubusercontent.com/render/math?math=10^{8}">` will produce <img src="https://render.githubusercontent.com/render/math?math=10^{8}">
