### 2020-07-16 #2432 @betteridiot
Thank you. Apparently it was just a browser extension that was causing the problem.
