### 2020-07-17 czbiohub/cerebra#89 @lincoln-harris
thats a great idea - yeah we've noticed that A) VCFs vary greatly in size depending on the experimental source and B) to state the obvious, larger VCFs run much, much slower.
