### 2020-07-19 czbiohub/cerebra#82 @afrubin
This looks much better. The "Pull Request Guidelines" section still needs some work, though. For example, it doesn't specify test coverage and README.rst doesn't exist (you have a README.md).
