### 2020-07-31 #2432 @lincoln-harris
the majority of reviewer comments have been addressed with the latest [release](https://github.com/czbiohub/cerebra/releases/tag/v1.1.2).
looking forward to additional feedback!
