### 2020-07-31 czbiohub/cerebra#83 @afrubin
The new installation section in the README is greatly improved, but still I have some comments and concerns.

You should mention the Windows limitations at the start of the installation section. It may actually be possible to run the software using docker or WSL (maybe you or one of your co-authors can test this?) and it's fair for you to state that these solutions exist but are not officially supported.

Instructions for users and developers are mixed together. I recommend that you split this into two suitably-titled subsections, possibly with a third subsection describing the additional dependencies required by all pip install methods.

The docker build seems to be working nicely and you might even want to suggest this as the recommended way to run the software, with pip installation instructions provided for those who don't want to or can't run docker.

User instructions for pip installation should not install the package from source in editable mode. Instead, this should have the commands for creating a venv or conda environment and then installing from pypi.

Developer instructions begin by reminding the reader to create their own fork and clone that. These instructions only work for you and other contributors to the main repo.

Instructions for system-wide pypi install should be removed, since it is almost always better to use a virtual environment and this creates a pitfall for novice users.

+1 for adding the classifiers to setup.py as suggested by @betteridiot. I should point out that there is a `Topic :: Scientific/Engineering :: Bio-Informatics` that might be a better fit than Information Technology.
