### 2020-07-31 czbiohub/cerebra#82 @afrubin
The Contributing section in the README mentions a CONTRIBUTING.md that no longer exists.
