### 2020-07-31 #2432 @afrubin
@lincoln-harris I have a few additional comments on the writing in the germline-filter section of the paper and the README.

A more typical phrasing from cancer genomics would be "tumor vs. normal" rather than "tumor/pathogenic vs control."

The text says that germline-filter can be used "so as to not bias the results by including non-pathogenic variants." While I understand the motivation for this, you can't say that germline variants are non-pathogenic - they may well be strongly pathogenic for another disease!

Instead of "control and the experimental tissue" the text should stick with the word "sample" (and tumor/normal) for clarity.
