### 2020-08-17 #2432 @betteridiot
Yes, tomorrow I will drop all my notes on the Issues page.
