### 2020-08-18 czbiohub/cerebra#105 @betteridiot
JoSS Review: Installation
With regard to the ongoing review for your [JoSS submission](https://github.com/openjournals/joss-reviews/issues/2432) here are my notes/questions:
1. Why suggest users to install via `pip` in editable (`-e`) mode?
2. For your `pip` installations you could write the command as: `pip install [--user] .` to free up any global install concerns
3. You should preface the OS dependencies _prior_ to the package installation process. This might prevent users from getting towards the end of the install process and throwing up their hands in defeat.

Below are my commands and history from following through the process.
```bash
$ git clone https://github.com/czbiohub/cerebra.git
Cloning into 'cerebra'...
remote: Enumerating objects: 20, done.
remote: Total 3239 (delta 5), reused 9 (delta 4), pack-reused 3219
Resolving deltas: 100% (2144/2144), done.
$ conda create -n cerebra python=3.7
Collecting package metadata (current_repodata.json): done
Solving environment: done

==> WARNING: A newer version of conda exists. <==
  current version: 4.8.3
  latest version: 4.8.4

Please update conda by running

    $ conda update -n base conda

## Package Plan ##

  environment location: /root/miniconda3/envs/cerebra

  added / updated specs:
    - conda
    - pip
    - python=3.7

The following packages will be downloaded:

    package                    |            build
    ---------------------------|-----------------
    _openmp_mutex-4.5          |            1_gnu          22 KB  conda-forge
    certifi-2020.6.20          |   py37hc8dfbb8_0         151 KB  conda-forge
    cffi-1.14.1                |   py37h2b28604_0         223 KB  conda-forge
    conda-4.8.4                |   py37hc8dfbb8_2         3.1 MB  conda-forge
    conda-package-handling-1.7.0|   py37h8f50634_4         2.0 MB  conda-forge
    cryptography-3.0           |   py37hb09aad4_0         633 KB  conda-forge
    libgcc-ng-9.3.0            |      h24d8f2e_15         7.8 MB  conda-forge
    libgomp-9.3.0              |      h24d8f2e_15         378 KB  conda-forge
    libstdcxx-ng-9.3.0         |      hdf63c60_15         4.0 MB  conda-forge
    ncurses-6.2                |       he1b5a44_1         993 KB  conda-forge
    pip-20.2.2                 |             py_0         1.1 MB  conda-forge
    python-3.7.8               |h6f2ec95_1_cpython        53.1 MB  conda-forge
    readline-8.0               |       he28a2e2_2         281 KB  conda-forge
    setuptools-49.6.0          |   py37hc8dfbb8_0         948 KB  conda-forge
    sqlite-3.33.0              |       h4cf870e_0         1.4 MB  conda-forge
    tqdm-4.48.2                |     pyh9f0ad1d_0          53 KB  conda-forge
    wheel-0.35.1               |     pyh9f0ad1d_0          29 KB  conda-forge
    zlib-1.2.11                |    h516909a_1007         106 KB  conda-forge
    ------------------------------------------------------------
                                           Total:        76.2 MB

The following NEW packages will be INSTALLED:

  _libgcc_mutex      conda-forge/linux-64::_libgcc_mutex-0.1-conda_forge
  _openmp_mutex      conda-forge/linux-64::_openmp_mutex-4.5-1_gnu
  brotlipy           conda-forge/linux-64::brotlipy-0.7.0-py37h8f50634_1000
  ca-certificates    conda-forge/linux-64::ca-certificates-2020.6.20-hecda079_0
  certifi            conda-forge/linux-64::certifi-2020.6.20-py37hc8dfbb8_0
  cffi               conda-forge/linux-64::cffi-1.14.1-py37h2b28604_0
  chardet            conda-forge/linux-64::chardet-3.0.4-py37hc8dfbb8_1006
  conda              conda-forge/linux-64::conda-4.8.4-py37hc8dfbb8_2
  conda-package-han~ conda-forge/linux-64::conda-package-handling-1.7.0-py37h8f50634_4
  cryptography       conda-forge/linux-64::cryptography-3.0-py37hb09aad4_0
  idna               conda-forge/noarch::idna-2.10-pyh9f0ad1d_0
  ld_impl_linux-64   conda-forge/linux-64::ld_impl_linux-64-2.34-hc38a660_9
  libffi             conda-forge/linux-64::libffi-3.2.1-he1b5a44_1007
  libgcc-ng          conda-forge/linux-64::libgcc-ng-9.3.0-h24d8f2e_15
  libgomp            conda-forge/linux-64::libgomp-9.3.0-h24d8f2e_15
  libstdcxx-ng       conda-forge/linux-64::libstdcxx-ng-9.3.0-hdf63c60_15
  ncurses            conda-forge/linux-64::ncurses-6.2-he1b5a44_1
  openssl            conda-forge/linux-64::openssl-1.1.1g-h516909a_1
  pip                conda-forge/noarch::pip-20.2.2-py_0
  pycosat            conda-forge/linux-64::pycosat-0.6.3-py37h8f50634_1004
  pycparser          conda-forge/noarch::pycparser-2.20-pyh9f0ad1d_2
  pyopenssl          conda-forge/noarch::pyopenssl-19.1.0-py_1
  pysocks            conda-forge/linux-64::pysocks-1.7.1-py37hc8dfbb8_1
  python             conda-forge/linux-64::python-3.7.8-h6f2ec95_1_cpython
  python_abi         conda-forge/linux-64::python_abi-3.7-1_cp37m
  readline           conda-forge/linux-64::readline-8.0-he28a2e2_2
  requests           conda-forge/noarch::requests-2.24.0-pyh9f0ad1d_0
  ruamel_yaml        conda-forge/linux-64::ruamel_yaml-0.15.80-py37h8f50634_1001
  setuptools         conda-forge/linux-64::setuptools-49.6.0-py37hc8dfbb8_0
  six                conda-forge/noarch::six-1.15.0-pyh9f0ad1d_0
  sqlite             conda-forge/linux-64::sqlite-3.33.0-h4cf870e_0
  tk                 conda-forge/linux-64::tk-8.6.10-hed695b0_0
  tqdm               conda-forge/noarch::tqdm-4.48.2-pyh9f0ad1d_0
  urllib3            conda-forge/noarch::urllib3-1.25.10-py_0
  wheel              conda-forge/noarch::wheel-0.35.1-pyh9f0ad1d_0
  xz                 conda-forge/linux-64::xz-5.2.5-h516909a_1
  yaml               conda-forge/linux-64::yaml-0.2.5-h516909a_0
  zlib               conda-forge/linux-64::zlib-1.2.11-h516909a_1007

Downloading and Extracting Packages
pip-20.2.2           | 1.1 MB    | ############################################################################################################################### | 100%
cryptography-3.0     | 633 KB    | ############################################################################################################################### | 100%
setuptools-49.6.0    | 948 KB    | ############################################################################################################################### | 100%
wheel-0.35.1         | 29 KB     | ############################################################################################################################### | 100%
libgcc-ng-9.3.0      | 7.8 MB    | ############################################################################################################################### | 100%
sqlite-3.33.0        | 1.4 MB    | ############################################################################################################################### | 100%
conda-package-handli | 2.0 MB    | ############################################################################################################################### | 100%
readline-8.0         | 281 KB    | ############################################################################################################################### | 100%
zlib-1.2.11          | 106 KB    | ############################################################################################################################### | 100%
ncurses-6.2          | 993 KB    | ############################################################################################################################### | 100%
tqdm-4.48.2          | 53 KB     | ############################################################################################################################### | 100%
libstdcxx-ng-9.3.0   | 4.0 MB    | ############################################################################################################################### | 100%
_openmp_mutex-4.5    | 22 KB     | ############################################################################################################################### | 100%
certifi-2020.6.20    | 151 KB    | ############################################################################################################################### | 100%
cffi-1.14.1          | 223 KB    | ############################################################################################################################### | 100%
libgomp-9.3.0        | 378 KB    | ############################################################################################################################### | 100%
python-3.7.8         | 53.1 MB   | ############################################################################################################################### | 100%
conda-4.8.4          | 3.1 MB    | ############################################################################################################################### | 100%
Preparing transaction: done
Verifying transaction: done
Executing transaction: done
#
# To activate this environment, use
#
#     $ conda activate cerebra
#
# To deactivate an active environment, use
#
#     $ conda deactivate

$ conda activate cerebra
$ cd cerebra/
$ pip install -e .
Obtaining file:///root/Downloads/cerebra
Collecting biopython
  Downloading biopython-1.77-cp37-cp37m-manylinux1_x86_64.whl (2.3 MB)
     |████████████████████████████████| 2.3 MB 1.8 MB/s
Collecting click
  Downloading click-7.1.2-py2.py3-none-any.whl (82 kB)
     |████████████████████████████████| 82 kB 1.3 MB/s
Collecting cython==0.29.13
  Downloading Cython-0.29.13-cp37-cp37m-manylinux1_x86_64.whl (2.1 MB)
     |████████████████████████████████| 2.1 MB 33.7 MB/s
Collecting hgvs
  Downloading hgvs-1.5.1-py2.py3-none-any.whl (98 kB)
     |████████████████████████████████| 98 kB 5.6 MB/s
Collecting ncls
  Downloading ncls-0.0.53.tar.gz (463 kB)
     |████████████████████████████████| 463 kB 36.7 MB/s
Collecting numpy
  Downloading numpy-1.19.1-cp37-cp37m-manylinux2010_x86_64.whl (14.5 MB)
     |████████████████████████████████| 14.5 MB 27.5 MB/s
Collecting pandas
  Downloading pandas-1.1.0-cp37-cp37m-manylinux1_x86_64.whl (10.5 MB)
     |████████████████████████████████| 10.5 MB 6.1 MB/s
Collecting pathos
  Downloading pathos-0.2.6.zip (219 kB)
     |████████████████████████████████| 219 kB 5.1 MB/s
Collecting pyfaidx
  Downloading pyfaidx-0.5.9.1.tar.gz (32 kB)
Requirement already satisfied: setuptools>=38.6.0 in /root/miniconda3/envs/cerebra/lib/python3.7/site-packages (from cerebra==1.1.1) (49.6.0.post20200814)
Requirement already satisfied: tqdm in /root/miniconda3/envs/cerebra/lib/python3.7/site-packages (from cerebra==1.1.1) (4.48.2)
Collecting vcfpy
  Downloading vcfpy-0.13.0.tar.gz (1.0 MB)
     |████████████████████████████████| 1.0 MB 27.0 MB/s
Collecting bioutils<1.0,>=0.4.0
  Downloading bioutils-0.5.2.post3-py2.py3-none-any.whl (256 kB)
     |████████████████████████████████| 256 kB 4.2 MB/s
Collecting biocommons.seqrepo<1.0
  Downloading biocommons.seqrepo-0.6.2-py2.py3-none-any.whl (31 kB)
Collecting ipython
  Downloading ipython-7.17.0-py3-none-any.whl (786 kB)
     |████████████████████████████████| 786 kB 40.8 MB/s
Collecting configparser>=3.3.0
  Downloading configparser-5.0.0-py3-none-any.whl (22 kB)
Collecting parsley
  Downloading Parsley-1.3-py2.py3-none-any.whl (88 kB)
     |████████████████████████████████| 88 kB 4.0 MB/s
Collecting psycopg2-binary
  Downloading psycopg2_binary-2.8.5-cp37-cp37m-manylinux1_x86_64.whl (2.9 MB)
     |████████████████████████████████| 2.9 MB 26.1 MB/s
Requirement already satisfied: six in /root/miniconda3/envs/cerebra/lib/python3.7/site-packages (from hgvs->cerebra==1.1.1) (1.15.0)
Collecting attrs>=17.4.0
  Using cached attrs-19.3.0-py2.py3-none-any.whl (39 kB)
Collecting python-dateutil>=2.7.3
  Downloading python_dateutil-2.8.1-py2.py3-none-any.whl (227 kB)
     |████████████████████████████████| 227 kB 6.0 MB/s
Collecting pytz>=2017.2
  Downloading pytz-2020.1-py2.py3-none-any.whl (510 kB)
     |████████████████████████████████| 510 kB 27.5 MB/s
Collecting ppft>=1.6.6.2
  Downloading ppft-1.6.6.2.zip (106 kB)
     |████████████████████████████████| 106 kB 5.4 MB/s
Collecting dill>=0.3.2
  Downloading dill-0.3.2.zip (177 kB)
     |████████████████████████████████| 177 kB 25.0 MB/s
Collecting pox>=0.2.8
  Downloading pox-0.2.8.zip (128 kB)
     |████████████████████████████████| 128 kB 30.6 MB/s
Collecting multiprocess>=0.70.10
  Downloading multiprocess-0.70.10.zip (2.4 MB)
     |████████████████████████████████| 2.4 MB 31.6 MB/s
Requirement already satisfied: requests in /root/miniconda3/envs/cerebra/lib/python3.7/site-packages (from bioutils<1.0,>=0.4.0->hgvs->cerebra==1.1.1) (2.24.0)
Collecting yoyo-migrations
  Downloading yoyo_migrations-7.2.0-py3-none-any.whl (36 kB)
Collecting requests-html
  Downloading requests_html-0.10.0-py3-none-any.whl (13 kB)
Collecting pysam
  Downloading pysam-0.16.0.1-cp37-cp37m-manylinux1_x86_64.whl (9.9 MB)
     |████████████████████████████████| 9.9 MB 10.0 MB/s
Collecting coloredlogs
  Downloading coloredlogs-14.0-py2.py3-none-any.whl (43 kB)
     |████████████████████████████████| 43 kB 1.8 MB/s
Collecting backcall
  Downloading backcall-0.2.0-py2.py3-none-any.whl (11 kB)
Collecting jedi>=0.10
  Downloading jedi-0.17.2-py2.py3-none-any.whl (1.4 MB)
     |████████████████████████████████| 1.4 MB 26.8 MB/s
Collecting pexpect; sys_platform != "win32"
  Downloading pexpect-4.8.0-py2.py3-none-any.whl (59 kB)
     |████████████████████████████████| 59 kB 3.6 MB/s
Collecting decorator
  Downloading decorator-4.4.2-py2.py3-none-any.whl (9.2 kB)
Collecting pygments
  Downloading Pygments-2.6.1-py3-none-any.whl (914 kB)
     |████████████████████████████████| 914 kB 38.2 MB/s
Collecting prompt-toolkit!=3.0.0,!=3.0.1,<3.1.0,>=2.0.0
  Downloading prompt_toolkit-3.0.6-py3-none-any.whl (354 kB)
     |████████████████████████████████| 354 kB 4.8 MB/s
Collecting traitlets>=4.2
  Using cached traitlets-4.3.3-py2.py3-none-any.whl (75 kB)
Collecting pickleshare
  Using cached pickleshare-0.7.5-py2.py3-none-any.whl (6.9 kB)
Requirement already satisfied: urllib3!=1.25.0,!=1.25.1,<1.26,>=1.21.1 in /root/miniconda3/envs/cerebra/lib/python3.7/site-packages (from requests->bioutils<1.0,>=0.4.0->hgvs->cerebra==1.1.1) (1.25.10)
Requirement already satisfied: certifi>=2017.4.17 in /root/miniconda3/envs/cerebra/lib/python3.7/site-packages (from requests->bioutils<1.0,>=0.4.0->hgvs->cerebra==1.1.1) (2020.6.20)
Requirement already satisfied: idna<3,>=2.5 in /root/miniconda3/envs/cerebra/lib/python3.7/site-packages (from requests->bioutils<1.0,>=0.4.0->hgvs->cerebra==1.1.1) (2.10)
Requirement already satisfied: chardet<4,>=3.0.2 in /root/miniconda3/envs/cerebra/lib/python3.7/site-packages (from requests->bioutils<1.0,>=0.4.0->hgvs->cerebra==1.1.1) (3.0.4)
Collecting tabulate
  Downloading tabulate-0.8.7-py3-none-any.whl (24 kB)
Collecting sqlparse
  Downloading sqlparse-0.3.1-py2.py3-none-any.whl (40 kB)
     |████████████████████████████████| 40 kB 3.6 MB/s
Collecting pyquery
  Downloading pyquery-1.4.1-py2.py3-none-any.whl (22 kB)
Collecting pyppeteer>=0.0.14
  Downloading pyppeteer-0.2.2-py3-none-any.whl (145 kB)
     |████████████████████████████████| 145 kB 22.9 MB/s
Collecting parse
  Downloading parse-1.16.0.tar.gz (30 kB)
Collecting w3lib
  Downloading w3lib-1.22.0-py2.py3-none-any.whl (20 kB)
Collecting fake-useragent
  Downloading fake-useragent-0.1.11.tar.gz (13 kB)
Collecting bs4
  Downloading bs4-0.0.1.tar.gz (1.1 kB)
Collecting humanfriendly>=7.1
  Downloading humanfriendly-8.2-py2.py3-none-any.whl (86 kB)
     |████████████████████████████████| 86 kB 7.8 MB/s
Collecting parso<0.8.0,>=0.7.0
  Downloading parso-0.7.1-py2.py3-none-any.whl (109 kB)
     |████████████████████████████████| 109 kB 36.0 MB/s
Collecting ptyprocess>=0.5
  Using cached ptyprocess-0.6.0-py2.py3-none-any.whl (39 kB)
Collecting wcwidth
  Downloading wcwidth-0.2.5-py2.py3-none-any.whl (30 kB)
Collecting ipython-genutils
  Using cached ipython_genutils-0.2.0-py2.py3-none-any.whl (26 kB)
Collecting cssselect>0.7.9
  Downloading cssselect-1.1.0-py2.py3-none-any.whl (16 kB)
Collecting lxml>=2.1
  Downloading lxml-4.5.2-cp37-cp37m-manylinux1_x86_64.whl (5.5 MB)
     |████████████████████████████████| 5.5 MB 35.7 MB/s
Collecting appdirs<2.0.0,>=1.4.3
  Downloading appdirs-1.4.4-py2.py3-none-any.whl (9.6 kB)
Collecting pyee<8.0.0,>=7.0.1
  Downloading pyee-7.0.2-py2.py3-none-any.whl (12 kB)
Collecting websockets<9.0,>=8.1
  Downloading websockets-8.1-cp37-cp37m-manylinux2010_x86_64.whl (79 kB)
     |████████████████████████████████| 79 kB 6.5 MB/s
Collecting beautifulsoup4
  Downloading beautifulsoup4-4.9.1-py3-none-any.whl (115 kB)
     |████████████████████████████████| 115 kB 35.1 MB/s
Collecting soupsieve>1.2
  Downloading soupsieve-2.0.1-py3-none-any.whl (32 kB)
Building wheels for collected packages: ncls, pathos, pyfaidx, vcfpy, ppft, dill, pox, multiprocess, parse, fake-useragent, bs4
  Building wheel for ncls (setup.py) ... done
  Created wheel for ncls: filename=ncls-0.0.53-cp37-cp37m-linux_x86_64.whl size=1813415 sha256=3b7450600946c3c040c124a71b741d48ae717d8ed2d7f3ce21724d2b3acfc055
  Stored in directory: /root/.cache/pip/wheels/b9/5a/6e/14e0fa19b2c9749376eb91aa956485565bd4a3e7d37ec438aa
  Building wheel for pathos (setup.py) ... done
  Created wheel for pathos: filename=pathos-0.2.6-py3-none-any.whl size=77675 sha256=29d2d541ccdf98754866d6ab3223dc2094f73d2f6b3208c2f634cff01c7d37f0
  Stored in directory: /root/.cache/pip/wheels/34/c0/d9/580a0af39a46bc51b0ffe4ad8bc09b4ad4a99865d64b3d379d
  Building wheel for pyfaidx (setup.py) ... done
  Created wheel for pyfaidx: filename=pyfaidx-0.5.9.1-py3-none-any.whl size=25118 sha256=55a25b9374c3954d5a4b1f8e561cfc98112c26d11d26804faf59cea959504604
  Stored in directory: /root/.cache/pip/wheels/d8/2b/eb/61b2b24f40d3d68f5a0ec6883aea7e0a6bbc4d4db0707fe2d3
  Building wheel for vcfpy (setup.py) ... done
  Created wheel for vcfpy: filename=vcfpy-0.13.0-py2.py3-none-any.whl size=34278 sha256=2dc6f9a6d886f5ad9c919c81c0d71c0bbde454950ec2f6f8a29ac91257187cfb
  Stored in directory: /root/.cache/pip/wheels/7b/f4/07/5136706296f1b804543a7db4e3943cda0a8ac581d9b7adb2ae
  Building wheel for ppft (setup.py) ... done
  Created wheel for ppft: filename=ppft-1.6.6.2-py3-none-any.whl size=64744 sha256=fa2552f98c3e5dc1a94caf5c6a03f62acaf0e62714c75cfedac5e1a38c885701
  Stored in directory: /root/.cache/pip/wheels/b0/6d/96/a352a1cb50c4a197c05ac0ac95c4eaeed92874b8737497af34
  Building wheel for dill (setup.py) ... done
  Created wheel for dill: filename=dill-0.3.2-py3-none-any.whl size=78912 sha256=7ddc729ce73326712ddfe2c8941a30350ff5be47b9132c91135551c9ef2da3d3
  Stored in directory: /root/.cache/pip/wheels/72/6b/d5/5548aa1b73b8c3d176ea13f9f92066b02e82141549d90e2100
  Building wheel for pox (setup.py) ... done
  Created wheel for pox: filename=pox-0.2.8-py3-none-any.whl size=28288 sha256=5a6acd38217865b6755fc9dee7fb2545a8d35f38a2fbc6ed0c236cd636aea807
  Stored in directory: /root/.cache/pip/wheels/3d/a8/3a/712ebffd829266bcb1be24c2617849960fd885aa5f81db07dd
  Building wheel for multiprocess (setup.py) ... done
  Created wheel for multiprocess: filename=multiprocess-0.70.10-py3-none-any.whl size=108187 sha256=38e7e9f535e621d9bb76a5879ae5747ac45b150f96d30af53dfe2016e2d34dcb
  Stored in directory: /root/.cache/pip/wheels/69/a0/aa/14379d16112299afa0b5f8464971648ea2a660a8f6aa1ca088
  Building wheel for parse (setup.py) ... done
  Created wheel for parse: filename=parse-1.16.0-py3-none-any.whl size=23996 sha256=5729511ef07faf0fa557ec1b296e9fcb40b533afda8bc3f701ead8e4536642a9
  Stored in directory: /root/.cache/pip/wheels/37/38/13/bc4f6e36a416d53588e5957b870d23bc76aa1710dc08719ee4
  Building wheel for fake-useragent (setup.py) ... done
  Created wheel for fake-useragent: filename=fake_useragent-0.1.11-py3-none-any.whl size=13486 sha256=9edc66760848d1676043770fbfce6a9830f771977890fd150a3a36b1c2e0c42a
  Stored in directory: /root/.cache/pip/wheels/ed/f7/62/50ab6c9a0b5567267ab76a9daa9d06315704209b2c5d032031
  Building wheel for bs4 (setup.py) ... done
[…]
