### 2020-08-18 czbiohub/cerebra#107 @betteridiot
JoSS Review: Automated Tests
With regard to the ongoing review for your [JoSS submission](https://github.com/openjournals/joss-reviews/issues/2432), here are my notes about the automated testing of the package:

Coverage appears to be fine. Tests can be somewhat slow, but that is understandable. I have nothing further to add here.

```bash
$ pytest --cov=. tests/
=============================================================================================== test session starts ================================================================================================
rootdir: /root/Downloads/cerebra
plugins: cov-2.10.1
collected 49 items

tests/test_commandline.py ...                                                                                                                                                                                [  6%]
tests/test_count_variants.py s.....                                                                                                                                                                          [ 18%]
tests/test_count_variants_adnl.py .....                                                                                                                                                                      [ 28%]
tests/test_find_peptide_variants_cosmic_setup.py .                                                                                                                                                           [ 30%]
tests/test_find_peptide_variants_exceptions.py ...                                                                                                                                                           [ 36%]
tests/test_find_peptide_variants_modules.py ...                                                                                                                                                              [ 42%]
tests/test_find_peptide_variants_output.py ..                                                                                                                                                                [ 46%]
tests/test_find_peptide_variants_runtime.py ...                                                                                                                                                              [ 53%]
tests/test_germline_filter.py ...                                                                                                                                                                            [ 59%]
tests/test_os_utils.py ...                                                                                                                                                                                   [ 65%]
tests/test_protein_variant_predictor.py ...                                                                                                                                                                  [ 71%]
tests/test_protein_variant_predictor_cosmic_setup.py .                                                                                                                                                       [ 73%]
tests/test_utils.py .....s.......                                                                                                                                                                            [100%]

================================================================================================= warnings summary =================================================================================================
/root/miniconda3/envs/cerebra/lib/python3.7/site-packages/hgvs/location.py:30
  /root/miniconda3/envs/cerebra/lib/python3.7/site-packages/hgvs/location.py:30: DeprecationWarning: The usage of `cmp` is deprecated and will be removed on or after 2021-06-01.  Please use `eq` and `order` instead.
    @attr.s(slots=True, repr=False, cmp=False)

/root/miniconda3/envs/cerebra/lib/python3.7/site-packages/hgvs/location.py:81
  /root/miniconda3/envs/cerebra/lib/python3.7/site-packages/hgvs/location.py:81: DeprecationWarning: The usage of `cmp` is deprecated and will be removed on or after 2021-06-01.  Please use `eq` and `order` instead.
    @attr.s(slots=True, repr=False, cmp=False)

/root/miniconda3/envs/cerebra/lib/python3.7/site-packages/hgvs/location.py:206
  /root/miniconda3/envs/cerebra/lib/python3.7/site-packages/hgvs/location.py:206: DeprecationWarning: The usage of `cmp` is deprecated and will be removed on or after 2021-06-01.  Please use `eq` and `order` instead.
    @attr.s(slots=True, repr=False, cmp=False)

/root/miniconda3/envs/cerebra/lib/python3.7/site-packages/Bio/Alphabet/__init__.py:27
  /root/miniconda3/envs/cerebra/lib/python3.7/site-packages/Bio/Alphabet/__init__.py:27: PendingDeprecationWarning: We intend to remove or replace Bio.Alphabet in 2020, ideally avoid using it explicitly in your code. Please get in touch if you will be adversely affected by this. https://github.com/biopython/biopython/issues/2046
    PendingDeprecationWarning,

-- Docs: https://docs.pytest.org/en/stable/warnings.html

----------- coverage: platform linux, python 3.7.8-final-0 -----------
Name                                                   Stmts   Miss  Cover
--------------------------------------------------------------------------
cerebra/__init__.py                                        3      0   100%
cerebra/commandline.py                                    15      2    87%
cerebra/count_variants.py                                 92      3    97%
cerebra/find_peptide_variants.py                         142      1    99%
cerebra/germline_filter.py                                61      2    97%
cerebra/os_utils.py                                       24      4    83%
cerebra/protein_variant_predictor.py                     138     10    93%
cerebra/utils.py                                         234     28    88%
setup.py                                                  12     12     0%
tests/test_commandline.py                                 17      0   100%
tests/test_count_variants.py                              85     26    69%
tests/test_count_variants_adnl.py                         75      2    97%
tests/test_find_peptide_variants_cosmic_setup.py          35      2    94%
tests/test_find_peptide_variants_exceptions.py            83      6    93%
tests/test_find_peptide_variants_modules.py               60      6    90%
tests/test_find_peptide_variants_output.py                33      1    97%
tests/test_find_peptide_variants_runtime.py               68      0   100%
tests/test_germline_filter.py                             58      1    98%
tests/test_os_utils.py                                    18      0   100%
tests/test_protein_variant_predictor.py                   48      1    98%
tests/test_protein_variant_predictor_cosmic_setup.py      49      2    96%
tests/test_utils.py                                      119      5    96%
--------------------------------------------------------------------------
TOTAL                                                   1469    114    92%

============================================================================== 47 passed, 2 skipped, 4 warnings in 370.78s (0:06:10) ===============================================================================
```
