### 2020-08-18 czbiohub/cerebra#108 @betteridiot
JoSS Review: Community Guidelines
With regard to the ongoing review for your [JoSS submission](https://github.com/openjournals/joss-reviews/issues/2432), here is a thought regarding your Community Guidelines requirement.

I see you have a [CONTRIBUTING.md](https://github.com/czbiohub/cerebra/blob/master/docs/CONTRIBUTING.md), but it also may go a long way to create some Issue/bug/feature templates. You can find the process for setting them up here: https://docs.github.com/en/github/building-a-strong-community/configuring-issue-templates-for-your-repository
