### 2020-08-18 #2432 @betteridiot
My notes have been posted @lpantano and @lincoln-harris. If they are addressed, then my review will be complete. Sorry for the delay.
