### 2020-08-24 czbiohub/cerebra#106 @lincoln-harris
#3 is a known issue (#99) having to do with the biocommons/hgvs package.
ill take another look at this
