### 2020-08-24 czbiohub/cerebra#105 @lincoln-harris
with regards to point #1 - I think the basic idea with editable mode is to allow users to make changes to the package without having to reinstall it every time. am I incorrect here?

basically `pip install -e . ` is a better idea if end users intend to edit the code themselves, i believe.

it looks like you do the same thing [here](https://github.com/betteridiot/bamnostic#traditional-github-clone)?
