### 2020-08-24 czbiohub/cerebra#105 @betteridiot
It is, but `bamnostic` has an API that I intend on users modifying for their own purposes. `cerebra` doesn't have the same API in place. Therefore, it stands to reason that you don't expect users to modify the code. Additonally, `cerebra` produces an end-point product of some analysis pipeline (even if it is initial dataset exploration). `bamnostic` fits somewhere in the middle of the pipeline and expects users to "use" the data during it.

What you are talking about is something like a `python setup.py develop` install instead.
