### 2020-08-24 czbiohub/cerebra#105 @betteridiot
Nonetheless, you are correct that I suggest the same thing for `bamnostic`. The difference is just in the intended use case: if you don't expect the users to want to (or need to) mess with the code, you should suggest it as such.

Think of it as what the intentionality of the code is: there are number of private methods and attributes within objects, but you don't necessarily document them. Why? Because you don't expect the end-user to _need_ to know what they are for and what they do. It would only serve to potentially confuse them.
