### 2020-08-24 czbiohub/cerebra#105 @lincoln-harris
ah i see -- youre totally right, `cerebra` has always been intended as an endpoint tool. in that light i agree it makes more sense for most users to install it in regular, rather than editable, mode.

out of curiosity can you link me to `bamnostic` code that implements the API? It might make sense to implement something similar with `cerebra` in the future.
