### 2020-08-24 czbiohub/cerebra#105 @betteridiot
Forgive me. Maybe API is now protected jargon and I should have chosen a different term.

By "API", I meant it in the most general sense: "application programming interface". It doesn't have a REST API or anything. However, much like `numpy` has a C API and `htslib` provides a C library for interfacing with sequencing data, `bamnostic` acts as the Python interface to sequencing data for the user.

That is, `bamnostic`'s functions and objects are written to be interacted with directly by the user. This is much like `numpy` or `pandas` is. A user would `import` my code and use it as necessary. Therefore, my functions must be written for the user to use. Not to be used by `bamnostic` alone.

For the sake of not digging myself any deeper: `bamnostic` is a library, not a standalone program.
