### 2020-08-25 czbiohub/cerebra#90 @lincoln-harris
answering your question (fourth bullet point): if the genome fasta file has already been indexed we do not reindex it; there should be a slight speed upgrade.
