### 2020-08-25 czbiohub/cerebra#106 @lincoln-harris
RE this cryptic error message:
```
<class 'hgvs.edit.AASub'> <class 'hgvs.edit.AARefAlt'>
```
your right, this was an error message meant for internal debugging.

shouldnt be seeing it anymore
