### 2020-08-25 czbiohub/cerebra#106 @lincoln-harris
2 and 4 should be addressed by the latest [PR](#109). basically the `report-coverage` option for `find-peptide-variants` returns a string that gives you some idea of the reads support for a given variant call. for example the string:
```
'ENSP00000395243.3:p.(Leu813delinsArgTrp),[2:0]'
```
tells you you have the ENSP00000395243.3 variant with an insertion of _ArgTrp_ in place of _Leu_ at position 813. At this loci we find 2 variant reads and 0 corresponding to the wildtype loci.
