### 2020-08-26 czbiohub/cerebra#106 @betteridiot
Since it is a boolean, maybe setting up `click` to use a [flag](https://click.palletsprojects.com/en/7.x/options/#boolean-flags) instead would be a more intuitive approach:

> Boolean flags are options that can be enabled or disabled. This can be accomplished by defining two flags in one go separated > by a slash (/) for enabling or disabling the option. (If a slash is in an option string, Click automatically knows that it’s a boolean > flag and will pass `is_flag=True` implicitly.) Click always wants you to provide an enable and disable flag so that you can change > the default later.
