### 2020-08-26 czbiohub/cerebra#83 @lincoln-harris
this is addressed with #109
