### 2020-08-26 #2432 @lincoln-harris
i believe everything has now been addressed.

most changes were made in [this PR](https://github.com/czbiohub/cerebra/pull/109), which I've since integrated into [`master`](https://github.com/czbiohub/cerebra/tree/master). that PR is annotated with a pretty thorough checklist of everything that I improved upon.

if I'm missing anything or there are any further comments, please let me know!
