### 2020-08-27 czbiohub/cerebra#83 @afrubin
I am still confused by the installation instructions for end users. Why do you recommend installing from source via git? It should be as simple as creating the virtual environment and using pip install to get the latest cerebra from PyPI.
