### 2020-09-03 #2432 @lpantano
@lincoln-harris, can you comment on the open issues opened by @betteridiot? if they are addressed or not?
