### 2020-09-14 #2432 @afrubin
Hi @lpantano, there are still a few open issues (such as https://github.com/czbiohub/cerebra/issues/83#issuecomment-681311968) and I feel like I should read the final version of the manuscript but I can't find a current PDF. I think it is very close though!
