### 2020-09-15 #2432 @lincoln-harris
sorry I've been away for a few weeks. I will work to address https://github.com/czbiohub/cerebra/issues/83#issuecomment-681311968 and do a better job of closing out issues that have been resolved.
