### 2020-09-15 czbiohub/cerebra#83 @lincoln-harris
ok so you think
```
conda create -n cerebra python=3.7
conda activate cerebra
pip install cerebra
```
is a better way of structuring it?
