### 2020-09-16 czbiohub/cerebra#83 @afrubin
Yes, thanks. These conda instructions and the venv instructions in the README are much better!
