### 2020-09-16 #2432 @afrubin
Looks good to me. My only remaining comment is that there's no API-level documentation for the code itself, and it looks like substantial extra work would be needed to generate the docstrings sphinx.ext.autodoc or similar tools would require.

However, given the focused nature of the software package (being an analysis tool not a library) and the detail in the usage guide, this might be OK. Happy to hear @lpantano weigh in on this.
