### 2020-09-16 #2432 @lpantano
Thanks everybody! I think they didn't aim the tool as a library, and I will feel ok not to mention the API. Although it could be a goal in the long term.

@afrubin, @betteridiot can you check all the items in the review list in the first comment of this review if you agree all has been addressed?

If they check all the items, @lincoln-harris, can you confirm the latest version, and prepare an archive for that version in zenodo where the authors list and title matches the paper and send me the DOI?

Thanks!
