### 2020-09-25 #2432 @lincoln-harris
ok the git archive has now been added to zenodo. DOI is 10.5281/zenodo.4050557, for latest version of the code, v1.2.0
