### 2020-09-29 #2432 @lincoln-harris
I dont think so? [paper.md](https://github.com/czbiohub/cerebra/blob/master/paper/paper.md) hasnt been changed since Aug 26, and it seemed to compile fine [Sept 15th](https://github.com/openjournals/joss-reviews/issues/2432#issuecomment-693142249)
