### 2020-10-01 #2432 @Kevin-Mattheus-Moerman
@lincoln-harris I am helping to process the acceptance of this work in JOSS. Can you work on the points below?

- [ ] You have one invalid DOI :point_up:, this is because you link to an arXiv preprint link rather than a valid DOI. Can you change the doi entry in the bib file to be of the `url` type instead? I believe this involves replacing:
```
 doi={arXiv:1207.3907v2}
```
with
```
 url={https://arxiv.org/abs/1207.3907}
```

- [ ] Please add a country to your affiliation(s) (please do not use acronyms for countries)

- [ ] The hyperlink to ENSEMBL in `Protein variants are converted to ENSEMBL protein IDs,...`, appears not to work, please check and update.
