### 2020-10-01 #2432 @Kevin-Mattheus-Moerman
@openjournals/dev I identified the invalid DOI but am unsure if this is the reason the paper does not compile for `@whedon accept`, can you confirm what the issue is here? Thanks
